Added checks for invalid filepaths for new file creation.

+++ DBFunctions.py
@@ -41,10 +41,20 @@ json_format = '''Please make sure data is formatted like so:
                     ]
                 }'''
 
+def json_file_exists(file_path):
+    try:
+        with open(file_path, "r") as file:
+            return True
+    except FileNotFoundError:
+        return False
+
 ############### READ/QUERY FUNCTIONS ###############
 
 # Function to get the JSON file to read from
 def json_read_file(file_path):
+    if json_file_exists(file_path) == False:
+        print("File does not exist. Please make sure the file path is correct.")
+        return None
     with open (file_path, "r") as file:
         data = json.load(file)
     return data
@@ -191,6 +201,9 @@ def json_get_parameter(data, class_name, method_name, parameter_name):
 
 # Function to write to a JSON file
 def json_write_file(file_path, data):
+    if json_file_exists(file_path) == False:
+        print("File does not exist. Please make sure the file path is correct.")
+        return None
     with open(file_path, "w") as file:
         json.dump(data, file, indent=4)
